Kyulacs

Documentation for Kyulacs.

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Base.Docs.Binding

Kyulacs.Circuit.QuantumCircuitOptimizer

Kyulacs.Gate.Adaptive

Kyulacs.Gate.AmplitudeDampingNoise

Kyulacs.Gate.BitFlipNoise

Kyulacs.Gate.CNOT

Kyulacs.Gate.CP

Kyulacs.Gate.CPTP

Kyulacs.Gate.CZ

Kyulacs.Gate.DenseMatrix

Kyulacs.Gate.DephasingNoise

Kyulacs.Gate.DepolarizingNoise

Kyulacs.Gate.DiagonalMatrix

Kyulacs.Gate.FREDKIN

Kyulacs.Gate.H

Kyulacs.Gate.Identity

Kyulacs.Gate.IndependentXZNoise

Kyulacs.Gate.Instrument

Kyulacs.Gate.Measurement

Kyulacs.Gate.P0

Kyulacs.Gate.P1

Kyulacs.Gate.ParametricPauliRotation

Kyulacs.Gate.ParametricRX

Kyulacs.Gate.ParametricRY

Kyulacs.Gate.ParametricRZ

Kyulacs.Gate.Pauli

Kyulacs.Gate.PauliRotation

Kyulacs.Gate.Probabilistic

Kyulacs.Gate.ProbabilisticInstrument

Kyulacs.Gate.RX

Kyulacs.Gate.RY

Kyulacs.Gate.RZ

Kyulacs.Gate.RandomUnitary

Kyulacs.Gate.ReversibleBoolean

Kyulacs.Gate.S

Kyulacs.Gate.SWAP

Kyulacs.Gate.Sdag

Kyulacs.Gate.SparseMatrix

Kyulacs.Gate.StateReflection

Kyulacs.Gate.T

Kyulacs.Gate.TOFFOLI

Kyulacs.Gate.Tdag

Kyulacs.Gate.TwoQubitDepolarizingNoise

Kyulacs.Gate.U1

Kyulacs.Gate.U2

Kyulacs.Gate.U3

Kyulacs.Gate.X

Kyulacs.Gate.Y

Kyulacs.Gate.Z

Kyulacs.Qulacs.DensityMatrix

Kyulacs.Qulacs.GeneralQuantumOperator

Kyulacs.Qulacs.Observable

Kyulacs.Qulacs.ParametricQuantumCircuit

Kyulacs.Qulacs.PauliOperator

Kyulacs.Qulacs.QuantumCircuit

Kyulacs.Qulacs.QuantumCircuitSimulator

Kyulacs.Qulacs.QuantumGateBase

Kyulacs.Qulacs.QuantumGateMatrix

Kyulacs.Qulacs.QuantumGate_SingleParameter

Kyulacs.Qulacs.QuantumState

Kyulacs.Qulacs.QuantumStateBase

Kyulacs.Qulacs.StateVector

Kyulacs.Gate.add

Kyulacs.Gate.merge

Kyulacs.Gate.sqrtX

Kyulacs.Gate.sqrtXdag

Kyulacs.Gate.sqrtY

Kyulacs.Gate.sqrtYdag

Kyulacs.Gate.to_matrix_gate

Kyulacs.ObservableFunctions.create_observable_from_openfermion_file

Kyulacs.ObservableFunctions.create_observable_from_openfermion_text

Kyulacs.ObservableFunctions.create_split_observable

Kyulacs.QuantumOperator.create_quantum_operator_from_openfermion_file

Kyulacs.QuantumOperator.create_quantum_operator_from_openfermion_text

Kyulacs.QuantumOperator.create_split_quantum_operator

Kyulacs.State.drop_qubit

Kyulacs.State.inner_product

Kyulacs.State.partial_trace

Kyulacs.State.permutate_qubit

Kyulacs.State.tensor_product

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Kyulacs.Qulacs.DensityMatrix — Type

DensityMatrix

Kyulacs.Qulacs.GeneralQuantumOperator — Type

GeneralQuantumOperator

Kyulacs.Qulacs.Observable — Type

Observable

Kyulacs.Qulacs.ParametricQuantumCircuit — Type

ParametricQuantumCircuit

Kyulacs.Qulacs.PauliOperator — Type

PauliOperator

Kyulacs.Qulacs.QuantumCircuit — Type

QuantumCircuit

Kyulacs.Qulacs.QuantumCircuitSimulator — Type

QuantumCircuitSimulator

Kyulacs.Qulacs.QuantumGateBase — Type

QuantumGateBase

Kyulacs.Qulacs.QuantumGateMatrix — Type

QuantumGateMatrix

Kyulacs.Qulacs.QuantumGate_SingleParameter — Type

QuantumGate_SingleParameter

Kyulacs.Qulacs.QuantumState — Type

QuantumState

Kyulacs.Qulacs.QuantumStateBase — Type

QuantumStateBase

Kyulacs.Qulacs.StateVector — Type

StateVector(arg0: int) -> qulacs_core.QuantumState

StateVector

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Kyulacs.Gate.Adaptive — Type

Adaptive(*args, **kwargs)
Overloaded function.

1. Adaptive(gate: qulacs_core.QuantumGateBase, condition: Callable[[List[int]], bool]) -> qulacs_core.QuantumGateBase

Create adaptive gate

2. Adaptive(gate: qulacs_core.QuantumGateBase, condition: Callable[[List[int], int], bool], id: int) -> qulacs_core.QuantumGateBase

Kyulacs.Gate.AmplitudeDampingNoise — Type

AmplitudeDampingNoise(index: int, prob: float) -> qulacs_core.QuantumGateBase

Create amplitude damping noise

Kyulacs.Gate.BitFlipNoise — Type

BitFlipNoise(index: int, prob: float) -> qulacs_core.QuantumGateBase

Create bit-flip noise

Kyulacs.Gate.CNOT — Type

CNOT(control: int, target: int) -> qulacs_core.QuantumGateBase

Create CNOT gate

Kyulacs.Gate.CP — Type

CP(kraus_list: List[qulacs_core.QuantumGateBase], state_normalize: bool, probability_normalize: bool, assign_zero_if_not_matched: bool) -> qulacs_core.QuantumGateBase

Create completely-positive map

Kyulacs.Gate.CPTP — Type

CPTP(kraus_list: List[qulacs_core.QuantumGateBase]) -> qulacs_core.QuantumGateBase

Create completely-positive trace preserving map

Kyulacs.Gate.CZ — Type

CZ(control: int, target: int) -> qulacs_core.QuantumGateBase

Create CZ gate

Kyulacs.Gate.DenseMatrix — Type

DenseMatrix(*args, **kwargs)
Overloaded function.

1. DenseMatrix(index: int, matrix: numpy.ndarray[numpy.complex128[m, n]]) -> qulacs_core.QuantumGateMatrix

Create dense matrix gate

2. DenseMatrix(index_list: List[int], matrix: numpy.ndarray[numpy.complex128[m, n]]) -> qulacs_core.QuantumGateMatrix

Kyulacs.Gate.DephasingNoise — Type

DephasingNoise(index: int, prob: float) -> qulacs_core.QuantumGateBase

Create dephasing noise

Kyulacs.Gate.DepolarizingNoise — Type

DepolarizingNoise(index: int, prob: float) -> qulacs_core.QuantumGateBase

Create depolarizing noise

Kyulacs.Gate.DiagonalMatrix — Type

DiagonalMatrix(index_list: List[int], diagonal_element: numpy.ndarray[numpy.complex128[m, 1]]) -> qulacs_core.QuantumGateBase

Create diagonal matrix gate

Kyulacs.Gate.FREDKIN — Type

FREDKIN(control: int, target1: int, target2: int) -> qulacs_core.QuantumGateMatrix

Create FREDKIN gate

Kyulacs.Gate.H — Type

H(index: int) -> qulacs_core.QuantumGateBase

Create Hadamard gate

Kyulacs.Gate.Identity — Type

Identity(index: int) -> qulacs_core.QuantumGateBase

Create identity gate

Kyulacs.Gate.IndependentXZNoise — Type

IndependentXZNoise(index: int, prob: float) -> qulacs_core.QuantumGateBase

Create independent XZ noise

Kyulacs.Gate.Instrument — Type

Instrument(kraus_list: List[qulacs_core.QuantumGateBase], register: int) -> qulacs_core.QuantumGateBase

Create instruments

Kyulacs.Gate.Measurement — Type

Measurement(index: int, register: int) -> qulacs_core.QuantumGateBase

Create measurement gate

Kyulacs.Gate.P0 — Type

P0(index: int) -> qulacs_core.QuantumGateBase

Create projection gate to |0> subspace

Kyulacs.Gate.P1 — Type

P1(index: int) -> qulacs_core.QuantumGateBase

Create projection gate to |1> subspace

Kyulacs.Gate.ParametricPauliRotation — Type

ParametricPauliRotation(index_list: List[int], pauli_ids: List[int], angle: float) -> qulacs_core.QuantumGate_SingleParameter

Create parametric multi-qubit Pauli rotation gate

Kyulacs.Gate.ParametricRX — Type

ParametricRX(index: int, angle: float) -> qulacs_core.QuantumGate_SingleParameter

Create parametric Pauli-X rotation gate

Kyulacs.Gate.ParametricRY — Type

ParametricRY(index: int, angle: float) -> qulacs_core.QuantumGate_SingleParameter

Create parametric Pauli-Y rotation gate

Kyulacs.Gate.ParametricRZ — Type

ParametricRZ(index: int, angle: float) -> qulacs_core.QuantumGate_SingleParameter

Create parametric Pauli-Z rotation gate

Kyulacs.Gate.Pauli — Type

Pauli(index_list: List[int], pauli_ids: List[int]) -> qulacs_core.QuantumGateBase

Create multi-qubit Pauli gate

Kyulacs.Gate.PauliRotation — Type

PauliRotation(index_list: List[int], pauli_ids: List[int], angle: float) -> qulacs_core.QuantumGateBase

Create multi-qubit Pauli rotation

Kyulacs.Gate.Probabilistic — Type

Probabilistic(prob_list: List[float], gate_list: List[qulacs_core.QuantumGateBase]) -> qulacs_core.QuantumGateBase

Create probabilistic gate

Kyulacs.Gate.ProbabilisticInstrument — Type

ProbabilisticInstrument(prob_list: List[float], gate_list: List[qulacs_core.QuantumGateBase], register: int) -> qulacs_core.QuantumGateBase

Create probabilistic instrument gate

Kyulacs.Gate.RX — Type

RX(index: int, angle: float) -> qulacs_core.QuantumGateBase

Create Pauli-X rotation gate

Kyulacs.Gate.RY — Type

RY(index: int, angle: float) -> qulacs_core.QuantumGateBase

Create Pauli-Y rotation gate

Kyulacs.Gate.RZ — Type

RZ(index: int, angle: float) -> qulacs_core.QuantumGateBase

Create Pauli-Z rotation gate

Kyulacs.Gate.RandomUnitary — Type

RandomUnitary(*args, **kwargs)
Overloaded function.

1. RandomUnitary(index_list: List[int]) -> qulacs_core.QuantumGateMatrix

Create random unitary gate

2. RandomUnitary(index_list: List[int], seed: int) -> qulacs_core.QuantumGateMatrix

Kyulacs.Gate.ReversibleBoolean — Type

ReversibleBoolean(index_list: List[int], func: Callable[[int, int], int]) -> qulacs_core.QuantumGateBase

Create reversible boolean gate

Kyulacs.Gate.S — Type

S(index: int) -> qulacs_core.QuantumGateBase

Create pi/4-phase gate

Kyulacs.Gate.SWAP — Type

SWAP(target1: int, target2: int) -> qulacs_core.QuantumGateBase

Create SWAP gate

Kyulacs.Gate.Sdag — Type

Sdag(index: int) -> qulacs_core.QuantumGateBase

Create adjoint of pi/4-phase gate

Kyulacs.Gate.SparseMatrix — Type

SparseMatrix(index_list: List[int], matrix: scipy.sparse.csc_matrix[numpy.complex128]) -> qulacs_core.QuantumGateBase

Create sparse matrix gate

Kyulacs.Gate.StateReflection — Type

StateReflection(state: qulacs_core.QuantumStateBase) -> qulacs_core.QuantumGateBase

Create state reflection gate

Kyulacs.Gate.T — Type

T(index: int) -> qulacs_core.QuantumGateBase

Create pi/8-phase gate

Kyulacs.Gate.TOFFOLI — Type

TOFFOLI(control1: int, control2: int, target: int) -> qulacs_core.QuantumGateMatrix

Create TOFFOLI gate

Kyulacs.Gate.Tdag — Type

Tdag(index: int) -> qulacs_core.QuantumGateBase

Create adjoint of pi/8-phase gate

Kyulacs.Gate.TwoQubitDepolarizingNoise — Type

TwoQubitDepolarizingNoise(index1: int, index2: int, prob: float) -> qulacs_core.QuantumGateBase

Create two-qubit depolarizing noise

Kyulacs.Gate.U1 — Type

U1(index: int, lambda: float) -> qulacs_core.QuantumGateBase

Create QASM U1 gate

Kyulacs.Gate.U2 — Type

U2(index: int, phi: float, lambda: float) -> qulacs_core.QuantumGateBase

Create QASM U2 gate

Kyulacs.Gate.U3 — Type

U3(index: int, theta: float, phi: float, lambda: float) -> qulacs_core.QuantumGateBase

Create QASM U3 gate

Kyulacs.Gate.X — Type

X(index: int) -> qulacs_core.QuantumGateBase

Create Pauli-X gate

Kyulacs.Gate.Y — Type

Y(index: int) -> qulacs_core.QuantumGateBase

Create Pauli-Y gate

Kyulacs.Gate.Z — Type

Z(index: int) -> qulacs_core.QuantumGateBase

Create Pauli-Z gate

Kyulacs.Gate.add — Method

add(*args, **kwargs)
Overloaded function.

1. add(gate1: qulacs_core.QuantumGateBase, gate2: qulacs_core.QuantumGateBase) -> qulacs_core.QuantumGateMatrix

Add quantum gate matrices

2. add(gate_list: List[qulacs_core.QuantumGateBase]) -> qulacs_core.QuantumGateMatrix

Add quantum gate matrices

Kyulacs.Gate.merge — Method

merge(*args, **kwargs)
Overloaded function.

1. merge(gate1: qulacs_core.QuantumGateBase, gate2: qulacs_core.QuantumGateBase) -> qulacs_core.QuantumGateMatrix

Merge two quantum gate or gate list

2. merge(gate_list: List[qulacs_core.QuantumGateBase]) -> qulacs_core.QuantumGateMatrix

Kyulacs.Gate.sqrtX — Method

sqrtX(index: int) -> qulacs_core.QuantumGateBase

Create pi/4 Pauli-X rotation gate

Kyulacs.Gate.sqrtXdag — Method

sqrtXdag(index: int) -> qulacs_core.QuantumGateBase

Create adjoint of pi/4 Pauli-X rotation gate

Kyulacs.Gate.sqrtY — Method

sqrtY(index: int) -> qulacs_core.QuantumGateBase

Create pi/4 Pauli-Y rotation gate

Kyulacs.Gate.sqrtYdag — Method

sqrtYdag(index: int) -> qulacs_core.QuantumGateBase

Create adjoint of pi/4 Pauli-Y rotation gate

Kyulacs.Gate.to_matrix_gate — Method

to_matrix_gate(gate: qulacs_core.QuantumGateBase) -> qulacs_core.QuantumGateMatrix

Convert named gate to matrix gate

Kyulacs.State.drop_qubit — Method

drop_qubit(state: qulacs_core.QuantumState, target: List[int], projection: List[int]) -> qulacs_core.QuantumState

Drop qubits from state

Kyulacs.State.inner_product — Method

inner_product(state_bra: qulacs_core.QuantumState, state_ket: qulacs_core.QuantumState) -> complex

Get inner product

Kyulacs.State.partial_trace — Method

partial_trace(*args, **kwargs)
Overloaded function.

1. partial_trace(state: qulacs_core.QuantumState, target_traceout: List[int]) -> qulacs_core.DensityMatrix

Take partial trace

2. partial_trace(state: qulacs_core.DensityMatrix, target_traceout: List[int]) -> qulacs_core.DensityMatrix

Kyulacs.State.permutate_qubit — Method

permutate_qubit(*args, **kwargs)
Overloaded function.

1. permutate_qubit(state: qulacs_core.QuantumState, order: List[int]) -> qulacs_core.QuantumState

Permutate qubits from state

2. permutate_qubit(state: qulacs_core.DensityMatrix, order: List[int]) -> qulacs_core.DensityMatrix

Kyulacs.State.tensor_product — Method

tensor_product(*args, **kwargs)
Overloaded function.

1. tensor_product(state_left: qulacs_core.QuantumState, state_right: qulacs_core.QuantumState) -> qulacs_core.QuantumState

Get tensor product of states

2. tensor_product(state_left: qulacs_core.DensityMatrix, state_right: qulacs_core.DensityMatrix) -> qulacs_core.DensityMatrix

Base.Docs.Binding — Method

Extract the docstring of a method in a given class julia> using Kyulacs julia> state = QuantumState(4) help?> state.setcomputationalbasis setcomputationalbasis(self: qulacs.QuantumState, index: int) -> None

Set state to computational basis

Kyulacs.Circuit.QuantumCircuitOptimizer — Type

QuantumCircuitOptimizer

Kyulacs.QuantumOperator.create_quantum_operator_from_openfermion_file — Method

create_quantum_operator_from_openfermion_file(arg0: str) -> qulacs_core.GeneralQuantumOperator

Kyulacs.QuantumOperator.create_quantum_operator_from_openfermion_text — Method

create_quantum_operator_from_openfermion_text(arg0: str) -> qulacs_core.GeneralQuantumOperator

Kyulacs.QuantumOperator.create_split_quantum_operator — Method

create_split_quantum_operator(arg0: str) -> Tuple[qulacs_core.GeneralQuantumOperator, qulacs_core.GeneralQuantumOperator]

Kyulacs.ObservableFunctions.create_observable_from_openfermion_file — Method

create_observable_from_openfermion_file(arg0: str) -> qulacs_core.Observable

Kyulacs.ObservableFunctions.create_observable_from_openfermion_text — Method

create_observable_from_openfermion_text(arg0: str) -> qulacs_core.Observable

Kyulacs.ObservableFunctions.create_split_observable — Method

create_split_observable(arg0: str) -> Tuple[qulacs_core.Observable, qulacs_core.Observable]